Create query and mutation types and resolvers
* Create typeDefs and define the types necessary for the queries.
* Create resolvers to handle the responses to the queries.
* Create a dataSources to hold the services need to make the database
requests.
* Separate the code into different files to organize the code.
* Modify account and transaction entities to have and return the
columns need when create a query. Those columns contain information
about other tables, like currency, exchange and user.
* Modify userId attribute to be undefined or a string in case it is not
give by the function call arguments.
* Add graphql optional attribute to handle when a request is from the
graphql api or the rest api.
* Create a scalar function to handle when a attribute is type Date
because this type is not supported natively in graphql.
* Clean the data return from the api when the call is from the rest
api. This is to maintain the rest api as it is and no populate it with
data not need.